issue fix

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('email');
             $table->string('status')->index()->comment('1-active,0-deactivated');
-            $table->tinyInteger('type')->index()->comment('1-Admin,2-Landlord,3-Tenant ');
+            $table->integer('type')->index()->comment('1-Admin,2-Landlord,3-Tenant ');
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->integer('landlord_id')->nullable();
